Rename ddc → ddc-services and ddc-local → ddc-project
